game APIを別ファイルに
